rename util.rs to room_resolver.rs
