+++ src/room_resolver.rs
